Add automatic chore assignment functionality
The new AutoAssign.py module handles automatic chore assignment.
Chores are assigned so as to even out the workload as much as possible.
Previous and future assigned work is taken into account in this
calculation.
Interacts with DataInput to access and change the chores database.

Calling the assign_unassigned_chores() function handles all of this.