Add chore repetition functionality
Adds the crucial function AutoAssign.renew_repeating_chores() along with
other functions in DataInput.py that make this possible.

This also brings a new chore status (RENEWED) and attribute (Frequency).

Upon renewal of a chore:
(1) The current chore is marked as RENEWED so it is never renewed again.
(2) A duplicate of that chore is created with status UNASSIGNED and a
new deadline calculated based on the frequency.