Clean up unused functions from DataInput.py
Removed functions from DataInput.py which did not have any usages. These
were mainly remnants from the very first work on this file, which
contained a lot of CLI elements which are now undesireable in a Python
file meant to be imported as a module.

I also reorganized everything under headers separating getter and setter
functions, which hopefully makes this a lot clearer to read (although
the git diff for this commit looks horrible)