Pass expected duration and deadline in chore add
Previously the expected duration would be set to a default value (10)
and the deadline would not be set at all on chore creation, leading to
really bad behavior in the automatic chore assignment.

This has now been rectified. The duration is set according to the user's
input and the deadline is set (for now) as today + frequency.